Add pathlib extensions
The `pathlib_extensions` module backports any new additions to the
`pathlib` module from the standard library for Python 3.6+ users who
are using an older version of `pathlib` that is bundled with their
standard library. For example, `pathlib.PurePath.with_stem()` is new in
Python 3.9, but `pathlib_extensions` allows users on Python 3.6 through
3.8 to use it too. As well, it will contain experimental features that
are not found in `pathlib`.