Add experimental features
Functionality to check whether a filesystem path is an image, an audio
or a video file.

In medical imaging research we do load (many) media files and want to
know what kind of media file it is (top-level media type). So we make
our life easier and implement functionality to easily check whether a
file is an image, an audio or a video file, without having to
know its extension beforehand.