Fix file recognition bug
Fixed bug where image, audio, or video files with an upper case file
extension are not recognized as such.

Before the fix, for example, imageA.bmp was recognized as image file, as
expected. However, imageB.BMP was not recognized due to the file
extension being in upper case, which was therefore not found in the list
of filename extensions for images (all lower case).